Move phpunit.xml to root folder.
Although not a best practive to lend a `phpunit.xml` file, this is a
library and I don't see people benefiting a lot from changing the
current configuration to other that may please them.

If this becomes an issue, we can always start ignoring `phpunit.xml` on
the VCS and distribute a `phpunit.xml-dist` so people can use it as a
base configuration file.